Enable the usage of Device Plugins
In certain situations, processing resources are made available
through Device Plugins,  either because they are structures to
be isolated, or because  they require  some special treatment.
In these cases,  Intel Power Manager is  not able to recognize
device plugins and treat them as CPU resources.

This modification aims to enable the use of device plugins as
processing resources, whenever this indication is convenient.
To do so,  the user must declare,  during the  application of
PowerConfig, which "customDevices" must  be considered during
the admission of the pod.

This modification also exempts the user from using the "CPU"
directive in resource requests whenever an entry of custom
Device is indicated.

Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>