Change CPUs accounting system
This  commit aims to  change  the  accounting system of CPUs. In some
cases,  the  number  of  CPUs obtained  through  the  runtime library
function is slightly less than the actual/real number of CPUs.

Due to the update of CPUs accounting system in the Power Optimization
Library, even if the CPUs are recognized by the library, they are not
accounted for when creating profiles.

This change introduces reading the number of CPUs online through sysfs
(/sys/devices/system/cpu/online), just like it's done in the library.

Signed-off-by: Eduardo Juliano Alberti <eduardo.alberti@windriver.com>