Container deletion from the workload
When deleting pods that use the resources of a given power profile,
the workload may be incorrectly updated, removing all existing
container entries. This error can be observed whenever containers
with the same name, but different pods, share resources of the
same workload.

This commit changes the container identification of a workload,
searching for its UID instead of its name, as well as including
workload information when admitting a powerpod.

Signed-off-by: Eduardo Juliano Alberti <eduardo.alberti@windriver.com>